Optimize type imports, better typing
