+++ .eslintrc.js
@@ -42,6 +42,13 @@ module.exports = {
     'import/no-default-export': 'off',
     'import/order': 'error',
     '@typescript-eslint/no-explicit-any': 'warn',
+    '@typescript-eslint/consistent-type-imports': [
+      'error',
+      {
+        prefer: 'type-imports',
+        disallowTypeAnnotations: false,
+      },
+    ],
   },
   settings: {
     react: {
@@ -67,6 +74,7 @@ module.exports = {
         'import/no-default-export': 'off',
         'import/order': 'off',
         '@typescript-eslint/no-explicit-any': 'off',
+        '@typescript-eslint/consistent-type-imports': ['off'],
       },
     },
     {
