+++ common/components/general/Device/Device.tsx
@@ -1,6 +1,7 @@
 // Found on https://stackoverflow.com/questions/59494037/how-to-detect-the-device-on-react-ssr-app-with-next-js
 
-import React, { ReactNode } from 'react';
+import type { ReactNode } from 'react';
+import React from 'react';
 import * as rdd from 'react-device-detect';
 
 interface DeviceProps {
