+++ common/types/dataPages.ts
@@ -1,0 +1,11 @@
+export type DataPage =
+  | 'overview'
+  | 'traveltimes'
+  | 'slowzones'
+  | 'headways'
+  | 'dwells'
+  | 'ridership'
+  | 'service'
+  | 'more';
+
+export type DataPageNames = Record<DataPage, string>;
