+++ modules/slowzones/api/slowzones.ts
@@ -1,4 +1,4 @@
-import { DayDelayTotals, SlowZoneResponse } from '../../../common/types/dataPoints';
+import type { DayDelayTotals, SlowZoneResponse } from '../../../common/types/dataPoints';
 
 export const fetchDelayTotals = (): Promise<DayDelayTotals[]> => {
   const url = new URL(`/static/slowzones/delay_totals.json`, window.location.origin);
