+++ package.json
@@ -29,6 +29,7 @@
   },
   "devDependencies": {
     "@tailwindcss/forms": "^0.5.3",
+    "@types/lodash": "^4.14.191",
     "@types/node": "^18.11.18",
     "@types/react": "18.0.27",
     "@typescript-eslint/eslint-plugin": "^5.50.0",
